fix: Schedule lsp location logic
nvim-navic updates the file context on a few events which happen to be
the same as the default events for updating the winbar in this plugin.
So what happens is that the `get_location` function will be called
before the context is updated. By scheduling the function call I fixed
this issue.

The assignment of winbar (the one before lsp location stuff) is
scheduled as well. This is because if it weren't, it would have had some
blinking effects on the winbar.