fix(basename): pass nil as icon to Entry constructor if not devicons_ok
Entry.icon is meant to be either a nil value or a table in which the
first element is the icon/symbol and a value bound to the `highlight`
key that is the highlight group of that icon. Therefore, before this
commit an empty table would be assigned as Entry.icon which caused the
issue #22.

References: #22, #23