fix(api)!: remove get_entries method from the API functions
This API does not work as it was described in the readme. Moreover,
making this work requires lots of refactoring. This plugin is not
intended to act as a backend for something else, so removing it from the
API list is the best choice.

BREAKING CHANGE: This API still exists but there's no guarantee that
it'll have the same signature in the future.

Fix #18